Fixed compilation problem on AMD64 (probably all 64 bit processors). This problem probably needs a more permanent solution. It seems the code sometimes assume an unsigned int has the same length as a pointer. The fix changes this to use unsigned long instead, so its not cleaner, it will break on platformes where sizeof(*) != sizeof(unsigned long)
git-svn-id: svn://svn.gna.org/svn/openinput/trunk@121 c88adb9f-b0f9-0310-9f7c-9d7fd975b7ca

+++ src/internal.h
@@ -126,7 +126,7 @@ oi_device *device_get(unsigned char index);
 
 void device_pumpall();
 
-unsigned int device_windowid(char *str,
+unsigned long device_windowid(char *str,
                              char tok);
 
 int device_destroy(unsigned char index);
